wip
Moving all the old stuff to an "old" folder. I'll be able to reference
it from there, like the vimrc and tmux conf.

I'm working on 2 new Python scripts. Python might not be the perfect
answer here, but it's a good cross platform solution, and it's portable
to a rasberry pi I might want with a good shell env.

One of these scripts will be for setting up and updating the baseline
env. It'll clone or update various directories and scripts, such as
oh-my-zsh, fzf, tmux, and others.

The other script will use some kind of syncing tool (gist, repo,
synctool) to upload/download new settings.